fix(portal): Bug #51374 join: Split into univention-debhelper
`univention-install-joinscript` is a development tool, which should be
placed in a low-level package and not included in a high-level package
like `univention-join`, which pulls in lots of other dependencies and
thus cannot be built early enough for other packages to use.

For example there is this dependency loop:

univention-join
 univention-directory-manager-tools
  python-univention-directory-manager
   univention-admin-diary-client
    univention-join

Create a new package `univention-debhelper` and move
`univention-install-joinscript` there.

Update the 3 users in UCS.

Bug #51374