Bug #52345: do not fetch portal.json multiple times
portal.json was fetched multiple times (up to 15 - 20 times) in various
situations. The original reason to fetch it multiple times is in edit mode so
after modifying we give the listener module enough time to write the new
entries so that we get the latest state. This was originally implemented in
git:46b7495702f7ccdd1bdf3404d6f42f5a0bcf4269. We now only do this logic if
we are really in edit mode and save/modify something.