linter: Exclude the linter Dockerfile from dockerfilelint
"dockerfilelint" does not seem to allow inline ignore flags or a similar
mechanism. This avoids a complaint about the intentional usage of the tag
"latest" for the base image of the linter.

Note that we have coverage also for this particular file due to the usage of
"hadolint".

Background details regarding the exclude pattern:

- The "pre-commit" documentation has a background regarding the pattern.
  Pointer: https://pre-commit.com/#regular-expressions

- The fragment "(?x)" is Python specific and activates the verbose mode.
  Pointer: https://docs.python.org/3/library/re.html#re.X

- Following the advise from the documentation of "pre-commit" ensures that we
  can keep the reason why to do an exclusion directly above the correct line.

Issue: #548