Add a long intro to a concept
Think this will usually not be needed, and should become over time part of a
process agreement within the dev team and the other stakeholders.

In this case I added extra background, since this is one of the first Epics to
work through in our team. Hope this will help all involved stakeholders to
understand the approach better and get more value out of the concept.