Enable serving of icons based on the relative URLs
There is a hidden issue behind this: The icons accumulate on the UCS host when
packages (Debian, App Center) are installed. The URLs are out of the directory
and relative.

This aspect cannot become part of any of the containers and will force us to
come up with a different handling.

Potentially:

- Debian Packages --> they have to provide them via the current mechanism. The
  link in the "portal.json" should become an absolute one though.

- App Center containers --> should serve the image themselves if possible, and
  fallback to the debian package mechanism if that's not reasonably possible.
  The link has to become an absolute URL somehow.

- Kubernetes containers --> The link in the entry has to be an absolute URL.
  Containers should serve the image themselves if possible.
  Fallback to some other storage otherwise (?)

Alternatives:

- A storage with an API and access control is added (e.g. S3 clone)

- Icons of any mechanism must be placed there

- The content is available via HTTP

Essentially this is what the current UCS system has by using apache as reverse
proxy and the file system as storage.