refactor(notification-api): Simplify settings setup
FastAPI does suggest to use the settings support classes of Pydantic. Those
automatically read environment variables and provide parsing / schema checking.

The FastAPI documentation does provide an example with encapsulating the access
to the one settings instance trough a function in combination with the lru_cache
decorator which might be useful in some occasions.

We don't have a use yet, so there is no justification for keeping this around.

Issue: #655