build: Move images back into the root folder
These images are installed via the Debian package, compare the files in the
directory "debian".

The install script is using them to set up entries into the LDAP directory via
the UDM. Serving the assets related to these entries must function in the same
way as for any other entry.

For the assets this means that they will sit in the correct place within the UCS
machine and be served out via the Apache httpd from the UCS machine. Any
environment like Kubernetes or App Center will have to ensure that the
respective parts of the URL path tree are forwarded correctly.

The frontend distribution itself (the container image) should not contain these
files at all.

Issue: #655