build(portal-server): Use HTTP backend in the container
This ensures that we use the HTTP based cache implementation within the
container. This configuration is put directly into the image, since that is at
the moment our only supported mode of operation when running as a container.

Technically it is still possible to bind mount a different configuration into
the container by using e.g. an adjusted configuration for docker compose.
Compare the notes in the compose file.

Issue: #655