refactor(notifications-api): Drop "/notifications/{id}/confirm"
This endpoint as probably been in the design of the API to support the
Announcement feature. During the implementation it became clear that
Announcements are an entirely different data model and also work through a
different mechanism. For the Notification feature alone there is no use of this
endpoint defined.