refactor(umc): adjust UMCP set/ HTTP requests
Before the UMC-servers were unified there was only one set/ endpoint.
This is still provided for backwards compatibility but additionally it
was split into multiple endpoints: set/password, set/locale,
set/user/preferences/. To save the redirection and be future compatible
when the single set/ is removed all existing calls are adjusted.

Bug #43633