refactor(portal-listener): Ensure all dependencies are always installed via pip
Found that we had the missing dependency of the library "requests" in the
container "portal-listener" because of the update to UCS 5.0.4 when switching to
"container-ucs-base".

This unveiled that we had a missing parameter in the call to "pip3 install"
which is addressed in this commit, so that we should be more robust regarding
future updates.

This also reverts commit 659c61a52d98fe02c0e9677625ad933ebe89f86c.