feat: Add support for "enable_xheaders"
The support for "X-Headers" is needed if the portal-server runs behind a proxy
server. In this case the header "X-Scheme" should provide the scheme of the
original request, e.g. "http" or "https". This ensures that URLs are constructed
with the correct external scheme.

Docs Tornado Framework: https://www.tornadoweb.org/en/stable/guide/running.html#running-behind-a-load-balancer