fix: Revert "fix(listener): Skip rebuild of group membership cache in init container"
This reverts commit 758e424067a417b18e4547cc7cf52e60e19938c4.

The change did introduce new obstacles because the portal would be empty until
the portal listener did have enough time to process all events of the bootstrap
jobs. Technically this would not be wrong, still, it does lead to a bad
experience when doing evaluation deployments on slower clusters.

A smarter handling of the cache rebuild will have to come once we did build the
transition into the provisioning system.