feat: Reload all logged in Portal browser tabs on logout
This new feature works by creating a new endpoint in the UMC '/logout-sse'.
This is a server-sent event compatabile endpoint that will send the client a
message with the content of 'logout' when the specific session ends.

On the frontend the Portal starts a SharedWorker. This SharedWorker will
connect to this endpoint once and when it receives the logout message will
send a message to all attached open Portals and tells them to refresh. Using
SharedWorkers here comes with the limitation that this feature will not work
on Safari < 16. In this case the feature will simply not be available.

Performance considerations: Every browser having an instance of the Portal
open will need a long lived TCP connection to Apache2. There is however
basically 0 network traffic in that connection. On initial connection and on
every additional automatic reconnect a simple HTTP message is sent. Apart
from that only the final 'logout' message is sent. This should be more
performant than polling.

This feature is enabled by default and can be disabled via UCR variable
'portal/reload-tabs-on-logout'.

Bug #57467
Issue #2362