feat: Remove optional sub-chart "postgresql"
We are dropping the optional bundled "postgresql" sub-chart on the level of the
component charts to keep things simple and to avoid inconsistencies in the
client access configuration.

The assumption is that evaluation deployments are usually done based on the
umbrella chart or some other mechanism which also prepares the database, so that
we see little value in still offering the bundling on component level.