feat(server): Log feature toggle configuration after loading the configuration files
This shall help with troubleshooting, so that the effective configuration is
visible in the logs. This allows to understand if e.g. in a Kubernetes based
deployment the intended configuration did arrive in the server's process and
help to understand better if an issue is on the server side or on the frontend
side.