refactor(portal-server): Move reading of environment variables into the main module
Long term we want to validate the configuration from all sources via a utility
like Pydantic and most likely also increase the support for environment
variables to be used as configuration.

Moving the logging configuration into the function signature of "setup_logger"
does avoid that the function does have to discover its configuration into
receiving it as an input parameter.