fix(portal-server): Make usage of the class UMCPortal work again
The class does have two implemented usage patterns:

- As a delegate in the regular "Portal" class.

- As a portal itself, the generated configuration can be found via the
  "univention-portal" script.

The second way of using it was not covered by any check and broken due to the
added features and backports from the past.

This adds minimal coverage to ensure that the code stays in sync with evolving
features and extends the UMCPortal implementation so that it is working again in
this use case.