feat(portal-consumer): Store CA certificate in target volume
Storing the certificate in the container locally has no side-effect since the
main container would not have access to the certificate.

Note: TLS support is incomplete, this commit does only adjust the existing code
so that it does store the data into the correct place. More work will have to be
done to fully support TLS with a custom CA certificate.