fix(notifications-api): Move DATABASE_URL into Deployment
The value depends on parameter substitution on the Kubernetes level. This means
that the value has to be put into the Deployment object itself for now.

A long term fix would put the fragments into environment variables and assemble
the URL inside of the application code.

Issue: univention/dev/internal/team-nubus#892